# docs/swagger.yaml
definitions:
...
  dto.TrojanItemOutput:
    properties:
      asset_id:
        type: integer
      asset_name:
        type: string
      create_at:
        type: string
      line:
        description: 1是容器、穿透, 2是主机,直通
        type: integer
      port_id:
        type: integer
      port_name:
        type: string
      real_server:
        type: string
    type: object
  dto.TrojanListOutput:
    properties:
      list:
        description: 列表
        items:
          $ref: '#/definitions/dto.TrojanItemOutput'
        type: array
      total:
        description: 总数
        type: integer
    type: object
...
  middleware.Response:
    properties:
      data:
        type: object
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: object
      trace_id:
        type: object
    type: object
info:
  contact: {}
  license: {}
paths:
...
  /trojan/echo:
    get:
      consumes:
      - application/json
      description: 木马通信
      operationId: /trojan/echo
      parameters:
      - description: 资产ID
        in: query
        name: id
        required: true
        type: integer
      - description: 资产名
        in: query
        name: name
        required: true
        type: integer
      - description: 线路ID
        in: query
        name: line
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: WS即时通
      tags:
      - 浮标管理
  /trojan/service/list:
    get:
      consumes:
      - application/json
      description: 获取在线木马服务列表
      operationId: /trojan/service/list
      parameters:
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 每页个数
        in: query
        name: limit
        required: true
        type: integer
      - description: 当前页数
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TrojanListOutput'
              type: object
      summary: 获取在线木马服务列表
      tags:
      - 浮标管理
...
swagger: "2.0"